AT-ReID-fast/requirements.txt
numpy<2
pillow
pyyaml
timm>=0.9
torch
torchvision
tqdm
